feat(phase-19/01): expand terminal-native-coding-agent into multi-file TypeScript project
