+++ phases/19-capstone-projects/01-terminal-native-coding-agent/code/ts/README.md
@@ -1,0 +1,36 @@
+# Capstone 19/01 — Terminal-Native Coding Agent (TypeScript)
+
+Multi-file TypeScript harness for the plan/act/observe loop described in
+`../docs/en.md`. Offline, deterministic, zero network calls.
+
+## Layout
+
+```
+src/
+  index.ts     entry point; runs a scripted demo and the eval, then exits 0
+  repl.ts      interactive command parser (run / eval / help / quit)
+  harness.ts   the plan-act-observe loop, wired through the hook bus
+  hooks.ts     eight-event hook bus plus a destructive-command guard
+  model.ts     scripted offline LLM that drives the demo
+  tools.ts     read_file + run_shell with zod-validated args
+  plan.ts     PlanState (todo rewrite) + Budget (turn / token / dollar ceilings)
+  eval.ts      tiny pass/fail counter across three offline tasks
+  types.ts     shared shape definitions
+tests/
+  harness.test.ts
+  tools.test.ts
+```
+
+## Run
+
+```
+npm install
+npm start                # runs the scripted demo + offline eval, exits 0
+npm start -- --repl      # opens the interactive harness REPL
+npm test                 # node --test runner via tsx
+npm run typecheck        # tsc --noEmit
+```
+
+The non-interactive `npm start` path asserts that the eval reports `passed=3
+failed=0` and that the scripted run converges to an all-done plan. Any drift
+fails the run.
