+++ phases/19-capstone-projects/01-terminal-native-coding-agent/code/ts/src/plan.ts
@@ -1,0 +1,59 @@
+import type { Status, TodoItem } from "./types.ts";
+
+export class PlanState {
+  goal: string;
+  items: TodoItem[];
+
+  constructor(goal: string) {
+    this.goal = goal;
+    this.items = [];
+  }
+
+  rewrite(items: TodoItem[]): void {
+    this.items = items;
+  }
+
+  summary(): string {
+    const mark: Record<Status, string> = {
+      pending: " ",
+      in_progress: ">",
+      done: "x",
+      failed: "!",
+    };
+    const lines = [`GOAL: ${this.goal}`];
+    for (const it of this.items) {
+      lines.push(`  [${mark[it.status]}] ${it.id}. ${it.description}`);
+    }
+    return lines.join("\n");
+  }
+}
+
+export class Budget {
+  maxTurns = 50;
+  maxTokens = 200_000;
+  maxDollars = 5.0;
+  turnsUsed = 0;
+  tokensUsed = 0;
+  dollarsUsed = 0;
+
+  step(tokens: number, dollars: number): void {
+    this.turnsUsed += 1;
+    this.tokensUsed += tokens;
+    this.dollarsUsed += dollars;
+  }
+
+  exceeded(): string | null {
+    if (this.turnsUsed >= this.maxTurns) return "turn_limit";
+    if (this.tokensUsed >= this.maxTokens) return "token_limit";
+    if (this.dollarsUsed >= this.maxDollars) return "dollar_limit";
+    return null;
+  }
+
+  snapshot(): { turnsUsed: number; tokensUsed: number; dollarsUsed: number } {
+    return {
+      turnsUsed: this.turnsUsed,
+      tokensUsed: this.tokensUsed,
+      dollarsUsed: this.dollarsUsed,
+    };
+  }
+}
