+++ phases/19-capstone-projects/01-terminal-native-coding-agent/code/ts/src/tools.ts
@@ -1,0 +1,35 @@
+import { readFileSync } from "node:fs";
+import * as path from "node:path";
+import { z } from "zod";
+import type { ToolArgs, ToolFn } from "./types.ts";
+
+export const TRUNCATE_BYTES = 4096;
+
+export const ReadFileArgs = z.object({ path: z.string().min(1) });
+export const RunShellArgs = z.object({ cmd: z.string().min(1) });
+
+export function toolReadFile(sandbox: string, args: ToolArgs): string {
+  const parsed = ReadFileArgs.parse(args);
+  const full = path.resolve(sandbox, parsed.path);
+  const root = path.resolve(sandbox);
+  if (!full.startsWith(root + path.sep) && full !== root) {
+    throw new Error("path escapes sandbox");
+  }
+  const data = readFileSync(full, "utf8");
+  return data.slice(0, TRUNCATE_BYTES);
+}
+
+export function toolRunShell(_sandbox: string, args: ToolArgs): string {
+  const parsed = RunShellArgs.parse(args);
+  const stub: Record<string, string> = {
+    ls: "README.md\nsrc\ntests",
+    "git status": "On branch agent/demo\nnothing to commit, working tree clean",
+  };
+  const out = stub[parsed.cmd] ?? `(stub) ran: ${parsed.cmd}`;
+  return `exit=0\n${out.slice(0, TRUNCATE_BYTES)}`;
+}
+
+export const TOOLS: Record<string, ToolFn> = {
+  read_file: toolReadFile,
+  run_shell: toolRunShell,
+};
